Don't attempt to support IVsDebuggableProjectCfg in IVsProjectFlavorCfg.get_CfgType.
This results in an InvalidCastException and seems to have done so for a long time.  Prior to VS2022 Preview 4, VS would swallow this exception and there seemed to be no observable detrimental effects.  However, starting with VS2022 Preview 4, VS changed their handling of these calls and no longer swallow the exception.  The exceptions bubbles up the stack and prevents the .wixproj from opening.